+++ sample/src/main/java/io/snabble/testapp/HomeFragment.java
@@ -1,6 +1,5 @@
 package io.snabble.testapp;
 
-import android.content.Intent;
 import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -8,7 +7,6 @@ import androidx.fragment.app.Fragment;
 import io.snabble.sdk.Project;
 import io.snabble.sdk.Snabble;
 import io.snabble.sdk.ui.SnabbleUI;
-import io.snabble.sdk.ui.receipts.ReceiptListView;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -42,14 +40,6 @@ public class HomeFragment extends Fragment {
             @Override
             public void onClick(View v) {
                 ((BaseActivity)getActivity()).showShoppingCart();
-            }
-        });
-
-        v.findViewById(R.id.receipt_list).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(requireActivity(), ReceiptListActivity.class);
-                startActivity(intent);
             }
         });